fix(logcollector): break circular import between sdcm.logcollector and sdcm.monitorstack
Deferred the import of sdcm.monitorstack.ui in logcollector.py to a
classmethod on GrafanaEntity, breaking the circular dependency that
prevented test collection on Python 3.14. The class attribute
base_grafana_dashboards is replaced by get_base_grafana_dashboards()
classmethod with the deferred import.

[SCT-399]
